track now uses globalMaker
